requirements.txt
Flask~=3.1.0
python-dotenv~=1.0.1
ipinfo==5.1.1
folium~=0.18.0
phonenumbers~=8.13.50
python_whois
dnspython
Flask-SQLAlchemy~=3.1.1
openai==1.53.0
tiktoken==0.8.0
selenium==4.26.1
webdriver-manager==4.0.2
